changes made in the code i used api recourses and used polices to the posts and comments so not any one can delete them

+++ app/Http/Controllers/CommentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\CommentResource;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use function Pest\Laravel\json;
@@ -11,7 +12,8 @@ class CommentController extends Controller
 {
     public function index()
     {
-       return Comment::with('user')->get();
+       $comment=Comment::with('user')->get();
+       return CommentResource::collection($comment);
 
     }
     public function create(Request $request)
@@ -19,39 +21,31 @@ class CommentController extends Controller
         $request->validate([
             'body'=>'required',
             'post_id'=>'required|numeric|exists:posts,id',
-            'user_id'=>'required|numeric|exists:users,id'
+
         ]);
 
-       $comment= Comment::create
-       (
-        [
+       $comment= $request->user()->comment()->create([
+           'body'=>$request->body,
+           'post_id'=>$request->post_id
+       ]);
 
-             'body'=>$request->body,
-             'user_id'=>$request->user_id,
-             'post_id'=>$request->post_id
-        ]
-       );
 
-       return response()->json([
-           'data'=>$comment,
-           'message'=>'Comment created'
-       ],201);
+       return new CommentResource($comment);
     }
     public function update(Request $request,Comment $comment)
     {
+        $this->authorize('update',$comment);
         $request->validate([
             'body'=>'required',
 
         ]);
         $comment->body=$request->body;
         $comment->save();
-        return response()->json([
-            'data'=>$comment,
-            'message'=>'Comment updated'
-        ],200);
+        return new CommentResource($comment);
     }
     public function delete(Comment $comment)
     {
+        $this->authorize('delete',$comment);
         $comment->delete();
         return response()->json(['message'=>'Comment deleted'],204);
     }
